# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transport/routes:
    get:
      operationId: listTransportRoutes
      tags: [transport]
      summary: List transport routes with their stops
      responses:
        "200":
          description: Routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportRoute"
    post:
      operationId: createTransportRoute
      tags: [transport]
      summary: Add a route (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportRouteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportRoute"
  /transport/routes/{id}:
    delete:
      operationId: deleteTransportRoute
      tags: [transport]
      summary: Remove a route (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /transport/routes/{id}/stops:
    post:
      operationId: createTransportStop
      tags: [transport]
      summary: Add a stop to a route (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportStopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportStop"
  /transport/stops/{id}:
    delete:
      operationId: deleteTransportStop
      tags: [transport]
      summary: Remove a stop (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /transport/assignments:
    get:
      operationId: listTransportAssignments
      tags: [transport]
      summary: List student transport assignments (own children only for student/parent)
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportAssignment"
    post:
      operationId: createTransportAssignment
      tags: [transport]
      summary: Assign a student to a route/stop (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportAssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportAssignment"
  /transport/assignments/{studentId}:
    delete:
      operationId: deleteTransportAssignment
      tags: [transport]
      summary: Remove a student's transport assignment (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TransportStop:
      type: object
      required: [id, routeId, name, order]
      properties:
        id:
          type: integer
        routeId:
          type: integer
        name:
          type: string
        order:
          type: integer
        pickupTime:
          type: ["string", "null"]
        dropTime:
          type: ["string", "null"]
    TransportStopInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        order:
          type: integer
        pickupTime:
          type: ["string", "null"]
        dropTime:
          type: ["string", "null"]
    TransportRoute:
      type: object
      required: [id, schoolId, name]
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        stops:
          type: array
          items:
            $ref: "#/components/schemas/TransportStop"
    TransportRouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    TransportAssignment:
      type: object
      required: [id, studentId, routeId, stopId]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        routeId:
          type: integer
        stopId:
          type: integer
        route:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
        stop:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            pickupTime:
              type: ["string", "null"]
            dropTime:
              type: ["string", "null"]
        student:
          type: object
          properties:
            id:
              type: integer
            rollNo:
              type: string
            user:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
    TransportAssignmentInput:
      type: object
      required: [studentId, routeId, stopId]
      properties:
        studentId:
          type: integer
        routeId:
          type: integer
        stopId:
          type: integer

    # ── Dashboard ─────────────────────────────────────────────────────────
...