# lib/api-zod/src/generated/types/transportRoute.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * School Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { TransportStop } from './transportStop';

export interface TransportRoute {
  id: number;
  schoolId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  stops?: TransportStop[];
}